+++ src/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- 3BV and openings calculation
+
 ## [0.1.0] - 2016-10-06
 ### Added
 - Minesweeper, Tile, and Stopwatch classes containing base game logic